Change layout of test module to avoid loading uninstalled packages
Some libraries are not installed for production builds, and thus
should not be referenced in the __init__.py
