+++ mtp_bank_admin/apps/bank_admin/tests/test_disbursements.py
@@ -9,7 +9,7 @@ from mtp_common.test_utils import silence_logger
 from openpyxl import load_workbook
 import responses
 
-from . import (
+from .utils import (
     NO_TRANSACTIONS, mock_list_prisons,
     get_test_disbursements, temp_file, api_url,
     mock_bank_holidays, ResponsesTestCase
